seems close to working version. Added some output for assignment. Needs work on scheduling- getting multiple departures in a row.

+++ ProcessList.h
@@ -11,7 +11,7 @@
 #include <iostream>
 #include <iomanip>
 
-extern const int LENGTH;
+extern const float LENGTH;
 
 class ProcessList{
 public:
